Refactor combine banner notifications queries
